HYNDVID-7, HYNDVID-43: Group based mapping of Users to Actions in place
NOTE: this commit modifies existing migrations. Drop and recreate your
development database.

Actions can be permitted on a Class or an Instance of a class.
Permissions can be granted via the cancan ability DSL or by creating and
associating entries in the Actions table with a Group via the Permission
table.